removed the generated js files
