+++ src/backend/Util/__tests__/normalizeData.test.ts
@@ -1,14 +1,14 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var _1 = require(".");
-var incomingData = {
+import { normalizeData } from '..';
+
+const incomingData = {
     id: 'yash_deviceid',
     type: 'testing',
     'locationCoordinate.x': 555.333,
     something: 'is up',
     goal: 'Incoming from CMX normalizer topic',
 };
-var config = {
+
+const config = {
     location_x: 'locationCoordinate.x',
     location_y: 'locationCoordinate.y',
     location_z: 'locationCoordinate.z',
@@ -21,9 +21,10 @@ var config = {
     last_updated: 'lastSeen',
     last_location_updated: 'lastSeen',
     id: 'id',
-    type: 'type'
+    type:'type',
 };
-var output = [
+
+const output = [
     {
         location_x: 555.333,
         location_y: undefined,
@@ -37,17 +38,19 @@ var output = [
         last_updated: undefined,
         last_location_updated: undefined,
         id: 'yash_deviceid',
-        type: 'testing',
+        type:'testing',
         custom_data: {
             something: 'is up',
             goal: 'Incoming from CMX normalizer topic',
         },
     },
 ];
+
 // let ans = normalizeData(incomingData, config);
 // console.log(ans);
-describe('Util', function () {
-    it('normalizeData', function () {
-        expect(_1.normalizeData(incomingData, config)).toEqual(output);
+
+describe('Util', () => {
+    it('normalizeData', () => {
+        expect(normalizeData(incomingData, config)).toEqual(output);
     });
 });
